Clean up

+++ inc/acfcs-actions.php
@@ -48,6 +48,15 @@
 	}
 	add_action( 'acfcs_after_success_import', 'acfcs_reimport_cities' );
 
+
+	/**
+	 * Save location as single meta values
+	 *
+	 * @param $value
+	 * @param $post_id
+	 *
+	 * @return void
+	 */
 	function acfcs_save_single( $value, $post_id ) {
 		if ( isset( $value[ 'save_single' ] ) && 1 == $value[ 'save_single' ] ) {
 			if ( ! empty( $value[ 'countryCode' ] ) ) {
@@ -59,12 +68,11 @@
 			if ( ! empty( $value[ 'cityName' ] ) ) {
 				update_post_meta( $post_id, 'acfcs_search_city', $value[ 'cityName' ] );
 			}
-		} else {
+		} elseif ( $post_id ) {
 			// remove meta
 			delete_post_meta( $post_id, 'acfcs_search_country' );
 			delete_post_meta( $post_id, 'acfcs_search_state' );
 			delete_post_meta( $post_id, 'acfcs_search_city' );
 		}
-		// error_log(print_r($value,true));
 	}
 	add_action( 'acfcs_store_meta', 'acfcs_save_single', 10, 2 );
